requirements.test.txt
-r requirements.txt

# for tests
codecov==2.1.12
coverage==7.0.0
pytest==7.2.0
pytest-cov==3.0.0
pytest-aiohttp==1.0.4
pytest-asyncio==0.20.2
pyyaml==6.0
